webapp/requirements.txt
Flask
scikit-learn
scikit-image
google-api-python-client
google-cloud-datastore
google-cloud-storage
grpcio
pillow